transaction_processor/requirements.txt
kafka-python-ng==2.2.2
clickhouse-connect==0.6.23
